refactor: split camera and alarm apps
