+++ raspberrypi_central/smart-camera/app/service_manager/run_camera.py
@@ -1,6 +1,6 @@
 from typing import List, Callable
 
-from camera.camera import Camera
+from camera.models import Camera
 from camera.camera_config import camera_config
 from camera_analyze.camera_analyzer import Consideration, CameraAnalyzer
 from camera.camera_factory import camera_factory
@@ -83,7 +83,7 @@ class RunSmartCamera(RunService):
 
         # TODO: see issue #78
         self._stream = self.video_stream(self._camera.process_frame, resolution=(
-            CAMERA_WIDTH, CAMERA_HEIGHT), framerate=1, pi_camera=False)
+            CAMERA_WIDTH, CAMERA_HEIGHT), framerate=1, pi_camera=True)
 
     def is_restart_necessary(self, data = None) -> bool:
         new_roi = roi_camera_from_args(data)
