+++ raspberrypi_central/webapp/app/Makefile
@@ -3,23 +3,40 @@ group := $(shell id -g)
 dc := USER_ID=$(user) GROUP_ID=$(group) docker-compose
 de := docker-compose exec
 
+dc-prod := $(dc) -f ../docker-compose.yml -f ../docker-compose.prod.yml
+
 web_image=web
 db_image=database
 
 manage := $(de) $(web_image) python manage.py
 
+.PHONY: stop-all
+stop-all:
+	docker stop $(docker ps -aq)
 
-.PHONY: up
-up:
+.PHONY: prod-up
+prod-up:
+	$(dc) -f ../docker-compose.yml -f ../docker-compose.prod.yml -d up
+
+.PHONY: dev-up
+dev-up:
 	$(dc) up
+
+.PHONY: monitoring
+monitoring:
+	$(dc-prod) up -d loki promtail grafana
 
 .PHONY: rabbit-services
 rabbit-services:
-	$(dc) up rabbit rabbit_worker
+	$(dc-prod) up -d rabbit rabbit_worker
 
 .PHONY: web-services
 web-services:
-	$(dc) up celery_beat web database mqtt python_process_mqtt
+	$(dc-prod) up -d database celery_beat web telegram_bot
+
+.PHONY: mqtt-services
+mqtt-services:
+	$(dc-prod) up -d mqtt python_process_mqtt
 
 .PHONY: build-docker
 build-docker:
@@ -46,7 +63,6 @@ manage:
 .PHONY: createapp
 createapp:
 	$(manage) startapp $(name)
-	sudo chown -R ${USER}:${USER} $(name)
 
 .PHONY: db-login
 db-login:
