+++ raspberrypi_central/webapp/app/alarm/admin.py
@@ -1,14 +1,6 @@
 from django.contrib import admin
 
-from .models.camera import CameraROI, CameraRectangleROI, CameraMotionDetectedPicture, CameraMotionDetected, \
-    CameraMotionDetectedBoundingBox
-from .models.alarm import AlarmStatus, AlarmSchedule
+from alarm.models import AlarmStatus, AlarmSchedule
 
 admin.site.register(AlarmStatus)
 admin.site.register(AlarmSchedule)
-
-admin.site.register(CameraMotionDetected)
-admin.site.register(CameraMotionDetectedPicture)
-admin.site.register(CameraRectangleROI)
-admin.site.register(CameraROI)
-admin.site.register(CameraMotionDetectedBoundingBox)
