+++ raspberrypi_central/webapp/app/alarm/forms.py
@@ -1,7 +1,6 @@
 from django import forms
-from django.forms import modelformset_factory
 
-from alarm.models import AlarmSchedule, CameraRectangleROI, AlarmStatus, CameraROI
+from alarm.models import AlarmSchedule, AlarmStatus
 
 
 class AlarmScheduleForm(forms.ModelForm):
@@ -10,26 +9,6 @@ class AlarmScheduleForm(forms.ModelForm):
         model = AlarmSchedule
         fields = '__all__'
 
-
-class CameraROIForm(forms.ModelForm):
-    class Meta:
-        model = CameraROI
-        exclude = ('define_picture',)
-
-
-class CameraROIUpdateForm(forms.ModelForm):
-    class Meta:
-        model = CameraROI
-        exclude = ('define_picture', 'device')
-
-
-class CameraRectangleROIForm(forms.ModelForm):
-
-    class Meta:
-        model = CameraRectangleROI
-        fields = '__all__'
-
-CameraRectangleROIFormSet = modelformset_factory(CameraRectangleROI, exclude=('camera_roi',))
 
 class AlarmStatusForm(forms.ModelForm):
     class Meta:
