--- raspberrypi_central/webapp/app/alarm/models/__init__.py
@@ -1,2 +1,0 @@
-from .alarm import AlarmStatus, AlarmSchedule
-from .camera import  CameraROI, CameraMotionDetected, CameraRectangleROI, CameraMotionDetectedPicture
