+++ raspberrypi_central/webapp/app/camera/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class CameraConfig(AppConfig):
+    name = 'camera'
