+++ raspberrypi_central/webapp/app/camera/models.py
