+++ raspberrypi_central/webapp/app/notification/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
 
-from .models.user_notifications import UserFreeCarrier, UserSetting, UserTelegramBotChatId
-from .models.notifications import SystemNotification
+from notification.models import UserFreeCarrier, UserTelegramBotChatId, UserSetting, SystemNotification
 
 admin.site.register(UserFreeCarrier)
 admin.site.register(UserTelegramBotChatId)
