+++ raspberrypi_central/webapp/app/notification/models.py
@@ -1,5 +1,47 @@
+import uuid
+
 from django.conf import settings
 from django.db import models
+
+from devices.models import Device
+
+"""
+This model is removed and might be used later.
+It is not used yet because we have models like CameraMotionDetectedPicture.
+But the design of this Model remains good.
+"""
+# class Attachment(models.Model):
+#     S3 = 0
+#
+#     TYPE_CHOICES = [
+#         (S3, 'Amazon s3')
+#     ]
+#
+#     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+#
+#     bucket_name = models.CharField(max_length=60)
+#     object_name = models.CharField(max_length=100)
+#     storage_type = models.IntegerField(choices=TYPE_CHOICES)
+#
+#     def __str__(self):
+#         return '{0}_{1}_{2}'.format(self.bucket_name, self.object_name, self.storage_type)
+
+
+class SeverityChoice(models.TextChoices):
+    LOW = 'low'
+    MODERATE = 'moderate'
+    HIGH = 'high'
+
+
+class SystemNotification(models.Model):
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+
+    message = models.TextField()
+    severity = models.CharField(max_length=60, choices=SeverityChoice.choices)
+
+    devices = models.ManyToManyField(Device, blank=True)
+
+    created_at = models.DateTimeField(auto_now_add=True)
 
 
 class UserFreeCarrier(models.Model):
