--- raspberrypi_central/webapp/app/notification/models/__init__.py
@@ -1,2 +1,0 @@
-from .user_notifications import UserSetting, UserFreeCarrier, UserTelegramBotChatId
-from .notifications import SeverityChoice, SystemNotification
