+++ raspberrypi_central/webapp/app/templates/camera/camera_motion_detected_detail.html
