+++ raspberrypi_central/webapp/app/templates/camera/camera_roi_form.html
