+++ raspberrypi_central/webapp/app/templates/camera/camera_roi_list.html
@@ -10,11 +10,11 @@
         <div class="box">
             <div class="box__body">
                 <h2>Alarm roi list</h2>
-                <a href="{% url 'alarm:camera_roi-add' %}">Add a new ROI to a camera</a>
+                <a href="{% url 'camera:camera_roi-add' %}">Add a new ROI to a camera</a>
                 <ul>
                     {% for roi in rois %}
                         <li>
-                            <a href="{% url 'alarm:camera_roi-edit' roi.pk %}">{{ roi }}</a></li>
+                            <a href="{% url 'camera:camera_roi-edit' roi.pk %}">{{ roi }}</a></li>
                     {% endfor %}
                 </ul>
             </div>
