+++ raspberrypi_central/webapp/app/templates/camera/cameraroi_confirm_delete.html
